refactor(cli): extract property commands to bin/commands/properties.js
Move `property-list`, `property-get`, `property-set`, and
`property-delete` into a dedicated module.
